[FAB-5720] Re-submit tx if re-validation passes
For the reason stated in FAB-5720, we shouldn't immediately
commit the message after re-validation, but re-submit it with
updated config sequence so that we don't risk forking the chain
due to non-determinism.

Note that this behavior is not compatible with pre-v1.1 orderers.
In order to support mixed cluster where pre-v1.1 and v1.1 orderer
co-exist, V1.1_Orderer_BugFixes flag is used to determine if we
should re-submit the re-validated tx. If this flag is not set,
the orderer is effectively running in compatibility mode.

Change-Id: Ibccf6a83a7b04be8b86611bb374de4f726856880
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>