[FAB-6072] Panic on incompatibilities
Although gracefully leaving a channel might be a better option in the
future, because this behavior is not compatible with v1.0.x gossip, and
because of the complexity of this operation, a good first pass for v1.1
is simply to ensure that all processes panic if incompatible
configuration is pushed to the channel.

This CR causes the orderer to check for capabilities during the config
validatiion process it already undertakes.  It also causes the orderer
to panic if a config block is committed which contains capabilities
unknown to the orderer.

Finally, on join channel and on config update, the peer checks the
capabilities and panics on incompatibility.

Change-Id: Ieeeaf7a1924c42e001a3a973873a4343f134a0ae
Signed-off-by: Binh Q. Nguyen <binhn@us.ibm.com>
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>