[FAB-6080] Fix unset mod_policy in channel create
For all channels created by the v1.0.x orderer, the Channel group's
mod_policy was left unset. Although the channel is functional, it is not
possible to add config elements to the channel level.

Although it would be nice to simply change this behavior in v1.1,
unfortunately, because the v1.0 orderers validate that the config
computed for the new channel is correct, based on this flawed generation
scheme, v1.1 orderers would accept the config as correct, but v1.0 would
not. This could allow the same channel name when referenced on different
orderers to have different backing blockchains and is not acceptable.

Therefore, this fix uses the newly developed capabilities flags to
determine whether the incompatible behavior should be enabled for an
orderer or not.

Change-Id: I498ac5891eddeef0bc81a928ab8af7fa1ff8d798
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>